Almost done commit, everything seems to be working, but we have to clean up and add tests.
