+++ src/config.yaml.example
@@ -1,0 +1,26 @@
+offsite_1:
+  description: 'Remote Backup Job 1'
+  remote:
+    host: '10.1.1.1'
+    user: 'admin'
+    paths:
+      - '/share/Datafile'
+      - '/share/Backupfiles'
+  local:
+    path: '/Volumes/offsite1'
+  rsync:
+    options: '--progress --recursive --archive --delete --verbose --stats --human-readable --copy-links'
+
+
+offsite_2:
+  description: 'Remote Backup Job 2'
+  remote:
+    host: '10.1.1.2'
+    user: 'admin'
+    paths:
+      - '/share/Datafile'
+      - '/share/Backupfiles'
+  local:
+    path: '/Volumes/offsite2'
+  rsync:
+    options: '--progress --recursive --archive --delete --verbose --stats --human-readable --copy-links'
