+++ src/exceptions.py
@@ -1,0 +1,18 @@
+class RsyncNotFoundException(Exception):
+    pass
+
+
+class RemoveSocketNotFoundException(Exception):
+    pass
+
+
+class PathNotFoundException(Exception):
+    pass
+
+
+class ConfigJobNotFoundException(Exception):
+    pass
+
+
+class PathValidationException(Exception):
+    pass
